refactor(backend): extract direct prompt turn contracts
- Move force-bound skill directive and Pointy inventory helpers into a focused prompt module.\n- Keep direct_prompts.py compatibility exports for existing callers/tests.\n- Preserve direct turn contract wording and tool discipline behavior.
